fix: periksa kehamilan section
