+++ app/dashboard/periksa-kehamilan/create/form-section/FaktorResiko.tsx
@@ -1,0 +1,777 @@
+import {
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from '@/components/ui/form';
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from '@/components/ui/popover';
+import { Input } from '@/components/ui/input';
+import { FormWrapper, Row, TitleSection } from '../create-form';
+import { CalendarIcon } from 'lucide-react';
+import { format } from 'date-fns';
+import { Button } from '@/components/ui/button';
+import { cn } from '@/lib/utils';
+import { Calendar } from '@/components/ui/calendar';
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '@/components/ui/select';
+import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
+
+export function MendeteksiFaktroResikoDanResikoTinggi({ form }: any) {
+  return (
+    <section className="_KUNJUNGAN_NIFAS space-y-4">
+      <TitleSection
+        title="Mendeteksi Faktor Resiko dan Resiko Tinggi"
+        subtitle="Masukkan data pasien"
+      />
+      <FormWrapper>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.umur"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Umur &lt; 20 tahun / &gt; 35 tahun
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.kpd"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">KPD</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.paritas"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">Paritas &gt; 4x</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.perdarahan"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">Perdarahan</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.spasing"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">Spasing &lt; 2 tahun</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.infeksi"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">Infeksi</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.bb"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  BB &lt; 38 kg / LILA &lt; 23,5 cm / penambahan BB &lt; 9 kg
+                  selama hamil
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.preeklamsi"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">Infeksi</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.tb"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  TB &lt; 145 cm, Scoliosis, Khiposis
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.hb"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">HB &lt; 8gr %</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.sakitKronis"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Sedang/pernah sakit kronis: TB, kel.Jantung hati, ginjal,
+                  psikologis, SLE, DM, Hipertensi
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.kelainanLetak"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Kelainan Letak (Sungsang/Lintang)
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.abortus"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Pernah abortus berulang, KET, Mola Hidati-dosa, APP, KPD, Bayi
+                  Cacat Kongenital
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.anakBesarHidramnion"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Anak Besar Hidramnion, Gemeli
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.sc"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Riwayat SC, Forcep V Ekstrasi
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.ancamanPrematur"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">Ancaman Prematur</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.hpp"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Riwayat HPP, Infeksi Nifas
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.infeksiDBD"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Infeksi (DBD, Thipoid, Sepsis)
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.bayiBesar"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Riwayat bayi besar &gt; 400 gr
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.distocia"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">Distocia</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.faktorResiko.hb"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">HB &lt; 11 gr %</FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="mendeteksiFaktorResikoDanResikoTinggi.resikoTinggi.terdapat2FaktorResiko"
+            render={({ field }) => (
+              <FormItem className="flex items-center justify-between">
+                <FormLabel className="text-md">
+                  Terdapat &gt; 2 Faktor Resiko
+                </FormLabel>
+                <FormControl>
+                  <RadioGroup
+                    onValueChange={field.onChange}
+                    defaultValue={field.value}
+                    className="flex h-10 items-center space-x-2"
+                  >
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="true" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Ya</FormLabel>
+                    </FormItem>
+                    <FormItem className="flex items-center space-x-1 space-y-0">
+                      <FormControl>
+                        <RadioGroupItem value="false" />
+                      </FormControl>
+                      <FormLabel className="font-normal">Tidak</FormLabel>
+                    </FormItem>
+                  </RadioGroup>
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+        <Row>
+          <FormField
+            control={form.control}
+            name="kunjunganNifas.resikoTinggi.ditemukanTanggal"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Ditemukan Tanggal</FormLabel>
+                <FormControl>
+                  <Input {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={form.control}
+            name="kunjunganNifas.resikoTinggi.jenisResiko"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Jenis Resiko</FormLabel>
+                <FormControl>
+                  <Input {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </Row>
+      </FormWrapper>
+    </section>
+  );
+}
