+++ components/ui/calendar.tsx
@@ -55,4 +55,51 @@ function Calendar({
 }
 Calendar.displayName = 'Calendar';
 
-export { Calendar };
+function CalendarLahir({
+  className,
+  classNames,
+  showOutsideDays = true,
+  ...props
+}: CalendarProps) {
+  return (
+    <DayPicker
+      captionLayout="dropdown-buttons"
+      // showOutsideDays={showOutsideDays}
+      // className={cn('p-1', className)}
+      // classNames={{
+      //   months: 'flex flex-col sm:flex-row space-y-4 sm:space-x-4 sm:space-y-0',
+      //   month: 'space-y-4',
+      //   caption: 'flex justify-center pt-1 relative items-center',
+      //   caption_label: 'text-sm font-semibold',
+      //   nav: 'space-x-1 flex items-center',
+      //   nav_button_previous: 'absolute left-1',
+      //   nav_button_next: 'absolute right-1',
+      //   table: 'w-full border-collapse space-y-1',
+      //   head_row: 'flex',
+      //   head_cell:
+      //     'text-black rounded-full md w-10 font-semibold text-[0.8rem]',
+      //   row: 'flex w-full mt-0',
+      //   cell: 'h-9 w-10 text-center text-sm p-0 relative rounded-full',
+      //   day: 'h-9 w-9 p-0 font-semibold aria-selected:opacity-100 hover:bg-[#D0E4FF] text-black rounded-full',
+      //   day_range_end: 'day-range-end',
+      //   day_selected:
+      //     'bg-blue-500 text-neutral-50 hover:bg-neutral-900 hover:text-neutral-50 focus:bg-blue-500 focus:text-neutral-50 dark:bg-neutral-50 dark:text-neutral-900 dark:hover:bg-neutral-50 dark:hover:text-neutral-900 dark:focus:bg-neutral-50 dark:focus:text-neutral-900',
+      //   day_today: 'text-black bg-blue-300',
+      //   day_outside: 'day-outside text-black/50',
+      //   day_disabled: 'text-neutral-500 opacity-50 dark:text-neutral-400',
+      //   day_range_middle:
+      //     'aria-selected:bg-neutral-100 aria-selected:text-neutral-900 dark:aria-selected:bg-neutral-800 dark:aria-selected:text-neutral-50',
+      //   day_hidden: 'invisible',
+      //   ...classNames,
+      // }}
+      // components={{
+      //   IconLeft: ({ ...props }) => <ChevronLeft className="h-4 w-4" />,
+      //   IconRight: ({ ...props }) => <ChevronRight className="h-4 w-4" />,
+      // }}
+      // {...props}
+    />
+  );
+}
+Calendar.displayName = 'CalendarLahir';
+
+export { Calendar, CalendarLahir };
